MOHAWK: LB - Cleanup ~4000 LoC in main LB (II).
Create livingbooks_constants.h to handle lots of kLBs.
Move LBItem respectively. Move LBSCriptEntry to itemscript.
Define header files: item,itemscript,itemtypes,aniation,page.
